+++ zeebe/qa/integration-tests/src/test/java/io/camunda/zeebe/it/gateway/CustomFilter.java
@@ -1,0 +1,37 @@
+/*
+ * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH under
+ * one or more contributor license agreements. See the NOTICE file distributed
+ * with this work for additional information regarding copyright ownership.
+ * Licensed under the Camunda License 1.0. You may not use this file
+ * except in compliance with the Camunda License 1.0.
+ */
+package io.camunda.zeebe.it.gateway;
+
+import jakarta.servlet.Filter;
+import jakarta.servlet.FilterChain;
+import jakarta.servlet.FilterConfig;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.ServletRequest;
+import jakarta.servlet.ServletResponse;
+
+public class CustomFilter implements Filter {
+
+  @Override
+  public void init(final FilterConfig filterConfig) throws ServletException {
+    Filter.super.init(filterConfig);
+  }
+
+  @Override
+  public void doFilter(
+      final ServletRequest servletRequest,
+      final ServletResponse servletResponse,
+      final FilterChain filterChain) {
+
+    throw new ArrayStoreException("I'm FILTERING!!!!");
+  }
+
+  @Override
+  public void destroy() {
+    Filter.super.destroy();
+  }
+}
