+++ zeebe/qa/util/src/main/java/io/camunda/zeebe/qa/util/cluster/TestCluster.java
@@ -258,7 +258,8 @@ public final class TestCluster implements CloseableSilently {
   public ZeebeClientBuilder newClientBuilder() {
     return ZeebeClient.newClientBuilder()
         .usePlaintext()
-        .grpcAddress(availableGateway().grpcAddress());
+        .grpcAddress(availableGateway().grpcAddress())
+        .restAddress(availableGateway().restAddress());
   }
 
   /**
